+++ public/app/plugins/datasource/loki/live_streams.ts
@@ -33,10 +33,10 @@ export class LiveStreams {
 
     const data = new CircularDataFrame({ capacity: target.size });
     data.addField({ name: 'labels', type: FieldType.other }); // The labels for each line
-    data.addField({ name: 'ts', type: FieldType.time, config: { displayName: 'Time' } });
-    data.addField({ name: 'line', type: FieldType.string }).labels = parseLabels(target.query);
+    data.addField({ name: 'Time', type: FieldType.time, config: {} });
+    data.addField({ name: 'Line', type: FieldType.string }).labels = parseLabels(target.query);
     data.addField({ name: 'id', type: FieldType.string });
-    data.addField({ name: 'tsNs', type: FieldType.time, config: { displayName: 'Time ns' } });
+    data.addField({ name: 'tsNs', type: FieldType.time, config: {} });
     data.meta = { ...data.meta, preferredVisualisationType: 'logs' };
     data.refId = target.refId;